Done Crud UoF

+++ FantasticForum/Tests/Repository/SqlRepositoryTests.cs
@@ -1,13 +1,18 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using System.Data;
+using System.Data.Entity.Infrastructure;
+using System.Data.Objects;
 using System.IO;
 using System.Linq.Expressions;
+using System.Runtime.Remoting.Contexts;
 using Models;
 using Mvc.Infrastructure.Abstract;
 using Mvc.Infrastructure.Concrete;
 using NUnit.Framework;
 using System.Linq;
+using Assert = NUnit.Framework.Assert;
 
 namespace Tests.Repository
 {
@@ -34,21 +39,33 @@ namespace Tests.Repository
         {
             sections = new List<Section>
                 {
-                    new Section {Title = "Sport", Topics = new Collection<Topic>
+                    new Section
                         {
-                            new Topic {Title = "Topic 1"},
-                            new Topic {Title = "Topic 2"},
-                        }},
-                    new Section {Title = "Life", Topics = new Collection<Topic>
+                            Title = "Sport",
+                            Topics = new Collection<Topic>
+                                {
+                                    new Topic {Title = "Topic 1"},
+                                    new Topic {Title = "Topic 2"}
+                                }
+                        },
+                    new Section
                         {
-                            new Topic {Title = "Topic 3"},
-                        }},
-                    new Section {Title = "News", Topics = new Collection<Topic>
+                            Title = "Life",
+                            Topics = new Collection<Topic>
+                                {
+                                    new Topic {Title = "Topic 3"},
+                                }
+                        },
+                    new Section
                         {
-                            new Topic {Title = "Topic 4"},
-                            new Topic {Title = "Topic 5"},
-                            new Topic {Title = "Topic 6"},
-                        }}
+                            Title = "News",
+                            Topics = new Collection<Topic>
+                                {
+                                    new Topic {Title = "Topic 4"},
+                                    new Topic {Title = "Topic 5"},
+                                    new Topic {Title = "Topic 6"},
+                                }
+                        }
                 };
             sections.ForEach(section => context.Sections.Add(section));
             context.SaveChanges();
@@ -86,7 +103,7 @@ namespace Tests.Repository
                                         Title = "Life",
                                         Topics = new Collection<Topic>
                                             {
-                                                new Topic {Title = "Topic 3"}
+                                                new Topic {Title = "Topic 3", SectionId = 2}
                                             }
                                     }
                             }
@@ -102,7 +119,7 @@ namespace Tests.Repository
                                         Topics = new Collection<Topic>
                                             {
                                                 new Topic {Title = "Topic 1"},
-                                                new Topic {Title = "Topic 2"},
+                                                new Topic {Title = "Topic 2"}
                                             }
                                     },
                                 new Section
@@ -133,15 +150,7 @@ namespace Tests.Repository
             var actualResult = repository.Get(filter);
 
             var actualList = actualResult.ToList();
-            Assert.That(actualList.Count, Is.EqualTo(expectedResult.Count));
-
-            for (var i = 0; i < actualList.Count; i++)
-            {
-                Assert.That(actualList[i].Title, Is.EqualTo(expectedResult[i].Title));
-
-                for (var j = 0; j < actualList[i].Topics.Count; j++)
-                    Assert.That(actualList[i].Topics[j].Title, Is.EqualTo(expectedResult[i].Topics[j].Title));
-            }
+            Assert.That(actualList, Is.EquivalentTo(expectedResult));
         }
 
         [Test]
@@ -160,7 +169,7 @@ namespace Tests.Repository
         [Test]
         public void UpdateTest()
         {
-            var id = context.Sections.First().Id;
+            var id = context.Sections.First(s => s.Title == "Sport").Id;
             var section = GetEntityById(id);
             Assert.That(section.Id, Is.EqualTo(id));
             Assert.That(section.Title, Is.EqualTo("Sport"));
@@ -175,12 +184,38 @@ namespace Tests.Repository
         }
 
         [Test]
+        public void UpdateWithConcurrencyExceptionTest()
+        {            
+            var id = context.Sections.First(s => s.Title == "Sport").Id;
+            var section1 = GetEntityById(id);
+            Assert.That(section1.Id, Is.EqualTo(id));
+            Assert.That(section1.Title, Is.EqualTo("Sport"));
+            var section2 = section1.Clone() as Section;
+
+            section1.Title = "Games";
+
+            repository.Update(section1);
+
+            section1 = GetEntityById(id);
+            Assert.That(section1.Id, Is.EqualTo(id));
+            Assert.That(section1.Title, Is.EqualTo("Games"));
+
+            section2.Title = "War";
+            
+            Assert.Throws<DbUpdateConcurrencyException>(() => repository.Update(section2));       
+        }
+
+        [Test]
         public void RemoveTest()
         {
-            var id = context.Sections.First().Id;
+            var id = context.Sections.First(s => s.Title == "Sport").Id;
             var section = repository.GetById(id);
             Assert.That(section.Id, Is.EqualTo(id));
             Assert.That(section.Title, Is.EqualTo("Sport"));
+
+            for (var i = section.Topics.Count - 1; i >= 0; i--)
+                context.Topics.Remove(section.Topics[i]);
+            context.SaveChanges();
             repository.Remove(section);
 
             section = repository.GetById(id);
@@ -189,13 +224,41 @@ namespace Tests.Repository
         }
 
         [Test]
+        public void RemoveWithConcurrencyExceptionTest()
+        {
+            var id = context.Sections.First(s => s.Title == "Sport").Id;
+            var section1 = GetEntityById(id);
+            Assert.That(section1.Id, Is.EqualTo(id));
+            Assert.That(section1.Title, Is.EqualTo("Sport"));
+            var section2 = section1.Clone() as Section;
+
+            section1.Title = "Games";
+
+            repository.Update(section1);
+
+            section1 = GetEntityById(id);
+            Assert.That(section1.Id, Is.EqualTo(id));
+            Assert.That(section1.Title, Is.EqualTo("Games"));
+
+            for (var i = section2.Topics.Count - 1; i >= 0; i--)
+                context.Topics.Remove(section2.Topics[i]);
+            context.SaveChanges();
+            Assert.Throws<DbUpdateConcurrencyException>(() => repository.Remove(section2));
+        }
+
+        [Test]
         public void RemoveByIdTest()
         {
-            var id = context.Sections.First().Id;
+            var id = context.Sections.First(s => s.Title == "Sport").Id;
             var section = repository.GetById(id);
             Assert.That(section.Id, Is.EqualTo(id));
             Assert.That(section.Title, Is.EqualTo("Sport"));
+
+            for (var i = section.Topics.Count - 1; i >= 0; i--)
+                context.Topics.Remove(section.Topics[i]);
+            context.SaveChanges();
             repository.Remove(id);
+            context.SaveChanges();
 
             section = repository.GetById(id);
 
@@ -210,10 +273,20 @@ namespace Tests.Repository
         [TearDown]
         public void TearDown()
         {
+            var objectContext = ((IObjectContextAdapter)context).ObjectContext;
+            objectContext.Refresh(RefreshMode.StoreWins,
+                                  objectContext.ObjectStateManager.GetObjectStateEntries(EntityState.Modified | EntityState.Deleted)
+                                               .Select(x => x.Entity));
+
             context.Sections.ToList().ForEach(section =>
                 {
-                    for (var i = section.Topics.Count - 1; i >= 0; i--)
-                        context.Topics.Remove(section.Topics[i]);
+                    if (section.Topics != null)
+                    {
+                        for (var i = section.Topics.Count - 1; i >= 0; i--)
+                            context.Topics.Remove(section.Topics[i]);
+                        context.SaveChanges();
+                    }
+
                     context.Sections.Remove(section);
                 });
             context.SaveChanges();
